Scaffolding for sqs consumer
