+++ src/Ditto.Sqs.Consumer/Ditto.Sqs.Consumer.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<?xml version="1.0" encoding="utf-8" standalone="yes"?>
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
@@ -11,20 +11,21 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="3.1.4"/>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.4"/>
-    <PackageReference Include="Serilog" Version="2.9.0"/>
-    <PackageReference Include="SerilogTimings" Version="2.3.0"/>
-    <PackageReference Include="Serilog.Sinks.Console" Version="3.1.1"/>
-    <PackageReference Include="Serilog.Settings.Configuration" Version="3.1.0"/>
-    <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="3.1.4"/>
-    <PackageReference Include="Serilog.Sinks.Seq" Version="4.0.0"/>
-    <PackageReference Include="Serilog.Sinks.Datadog.Logs" Version="0.3.1"/>
-    <PackageReference Include="prometheus-net" Version="3.5.0"/>
-    <PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.3.101"/>
-    <PackageReference Include="AWSSDK.Kinesis" Version="3.3.100.203"/>
+    <PackageReference Include="Gateway.Extensions.Sqs.Consumers" Version="1.9.1" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="3.1.4" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.4" />
+    <PackageReference Include="Serilog" Version="2.9.0" />
+    <PackageReference Include="SerilogTimings" Version="2.3.0" />
+    <PackageReference Include="Serilog.Sinks.Console" Version="3.1.1" />
+    <PackageReference Include="Serilog.Settings.Configuration" Version="3.1.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="3.1.4" />
+    <PackageReference Include="Serilog.Sinks.Seq" Version="4.0.0" />
+    <PackageReference Include="Serilog.Sinks.Datadog.Logs" Version="0.3.1" />
+    <PackageReference Include="prometheus-net" Version="3.5.0" />
+    <PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.3.101" />
+    <PackageReference Include="AWSSDK.Kinesis" Version="3.3.100.203" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Ditto.Core\Ditto.Core.csproj"/>
+    <ProjectReference Include="..\Ditto.Core\Ditto.Core.csproj" />
   </ItemGroup>
 </Project>
