+++ src/Ditto.Sqs.Consumer/IEventStore.cs
@@ -1,0 +1,10 @@
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Ditto.Sqs.Consumer
+{
+    public interface IEventStore
+    {
+        Task Save(object a, CancellationToken c);
+    }
+}
