--- src/Ditto.Sqs.Consumer/KinesisSettings.cs
@@ -1,7 +1,0 @@
-namespace Ditto.Kinesis
-{
-    public class KinesisSettings
-    {
-        public string StreamName { get; set; }
-    }
-}
